refactor: 스케줄러 파일 구조 변경
