+++ src/main/java/be15fintomatokatchupbe/oauth/command/application/controller/YoutubeCommandController.java
@@ -4,7 +4,7 @@ import be15fintomatokatchupbe.common.dto.ApiResponse;
 import be15fintomatokatchupbe.common.exception.BusinessException;
 import be15fintomatokatchupbe.config.security.model.CustomUserDetail;
 import be15fintomatokatchupbe.oauth.command.application.Service.YoutubeCommandService;
-import be15fintomatokatchupbe.oauth.command.application.scheduler.YoutubeStatsSnapshotScheduler;
+import be15fintomatokatchupbe.oauth.scheduler.YoutubeStatsSnapshotScheduler;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpHeaders;
