+++ src/main/java/be15fintomatokatchupbe/oauth/scheduler/YoutubeStatsSnapshotScheduler.java
@@ -1,4 +1,4 @@
-package be15fintomatokatchupbe.oauth.command.application.scheduler;
+package be15fintomatokatchupbe.oauth.scheduler;
 
 import be15fintomatokatchupbe.common.exception.BusinessException;
 import be15fintomatokatchupbe.influencer.command.domain.aggregate.entity.Youtube;
